test: remove useless changes

+++ test/TestCommonHeapOrdering.t.sol
@@ -12,7 +12,7 @@ abstract contract TestCommonHeapOrdering is Test {
 
     address[] public accounts;
     uint256 public NB_ACCOUNTS = 50;
-    uint128 public MAX_SORTED_USERS = 50;
+    uint256 public MAX_SORTED_USERS = 50;
     address public ADDR_ZERO = address(0);
 
     function update(
@@ -296,7 +296,7 @@ abstract contract TestCommonHeapOrdering is Test {
 
         for (uint256 i = 10; i < 15; i++) update(accounts[i], 0, i - 9);
 
-        for (uint128 i = 10; i < 15; i++) assertLe(heap.accountsValue(i), 10);
+        for (uint256 i = 10; i < 15; i++) assertLe(heap.accountsValue(i), 10);
     }
 
     function testDecreaseIndexChanges() public {
